requirements.txt
blessed==1.20.0
boto3==1.34.125
botocore==1.34.125
cachetools==5.3.3
certifi==2024.6.2
chalice==1.31.1
charset-normalizer==3.3.2
click==8.1.7
google-auth==2.29.0
idna==3.7
inquirer==2.10.1
jmespath==1.0.1
kubernetes==29.0.0
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-editor==1.0.4
PyYAML==6.0.1
readchar==4.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
